+++ components/Details.js
@@ -13,8 +13,8 @@ import { ListItem, Button  } from "react-native-elements";
 import { Root, Toast, Fab, Separator } from "native-base";
 import Icon from "react-native-vector-icons/FontAwesome";
 import * as WebBrowser from "expo-web-browser";
-import Pokeball from "./assets/pokeball.png";
-import PokeballSprite from "./assets/pokeballSprite.png";
+import Pokeball from "../assets/pokeball.png";
+import PokeballSprite from "../assets/pokeballSprite.png";
 import Stats from "./Stats";
 import TypeBadges from "./TypeBadges";
 import Popover from "./Popover";
