+++ components/GenerationView.js
