+++ components/TypeBadges.js
